Fix pointer-to-int cast in cthread_exit: use (int)(uintptr_t)
exit() takes int; result is void *. Casting pointer directly to int
is undefined behavior. Cast through uintptr_t first to safely convert
the pointer value to an integer before truncating to int.

Closes #26

+++ osfmk/src/mach_services/lib/libcthreads/cthreads.c
@@ -1225,7 +1225,7 @@ cthread_exit(void *result)
 	    cthread_status.exit_thread = p;
 	    cthread_block(p, (cthread_fn_t) cthread_exit, result, CTHREAD_NULL);
 	}
-		exit((int) result);
+		exit((int)(uintptr_t) result);
     } 
 
     /*
